db_manager removed. from_id types moved to scientific.from_id
