+++ obi_one/scientific/from_id/cell_morphology_from_id.py
