+++ obi_one/scientific/tasks/ion_channel_modeling.py
@@ -13,7 +13,7 @@ from obi_one.core.block import Block
 from obi_one.core.scan_config import ScanConfig
 from obi_one.core.single import SingleConfigMixin
 from obi_one.core.task import Task
-from obi_one.database.ion_channel_recording_from_id import IonChannelRecordingFromID
+from obi_one.scientific.from_id.ion_channel_recording_from_id import IonChannelRecordingFromID
 from obi_one.scientific.blocks import ion_channel_equations as equations_module
 
 """
