+++ refactor_notes.txt
@@ -9,6 +9,6 @@ Todo:
 - Update all project notebooks
 - Move ninja files from tasks
 - Rename Contribute configs + tasks (speak with Dan)
-- Remove database manager. Only used in entitysdk examples
+- Fix entitysdk examples
 - Use obi-notebooks? Separate tickets?
 - Remove/use validations/cell_morphology_validation
